requirements.txt
flask==0.12.2
redis==2.10.6
kafka-python==1.3.4